move ListenerFunc to _test.go file to exclude form export
